Add test cases for components: SecureLogger, AESCredentials
Conform ChiperError, CredentialsError, StorageError to LocalizedError

+++ Sources/UDFEncryptedFileLogger/Encription/AES.swift
@@ -11,7 +11,7 @@ import Foundation
 enum AESCipher {}
 
 protocol StreamCipherable {
-  var blockSize: UInt64 { get }
+  var blockSize: Int { get }
   func encode(data: Data) throws -> Data
   func decode(data: Data) throws -> Data
   func finish() throws -> Data
